# [TextFormatter] Speed up (~40%). Fprintf is changed to buffer.Write*