--- a/formatter_bench_test.go
+++ b/formatter_bench_test.go
@@ -1,6 +1,7 @@
 package logrus
 
 import (
+	"fmt"
 	"testing"
 	"time"
 )
@@ -43,6 +44,15 @@ var largeFields = Fields{
 	"this":      "will",
 	"make":      "thirty",
 	"entries":   "yeah",
+}
+
+var errorFields = Fields{
+	"foo": fmt.Errorf("bar"),
+	"baz": fmt.Errorf("qux"),
+}
+
+func BenchmarkErrorTextFormatter(b *testing.B) {
+	doBenchmark(b, &TextFormatter{DisableColors: true}, errorFields)
 }
 
 func BenchmarkSmallTextFormatter(b *testing.B) {