As Fprintf is slower than buffer.WriteString, it's replaced to faster call.

Before:
```
BenchmarkSmallTextFormatter	  300000	      5841 ns/op	  14.89 MB/s
BenchmarkLargeTextFormatter	   50000	     24215 ns/op	  11.85 MB/s
BenchmarkSmallColoredTextFormatter	  300000	      5071 ns/op	  27.02 MB/s
BenchmarkLargeColoredTextFormatter	   50000	     27192 ns/op	  20.34 MB/s
```
Now:
```
BenchmarkSmallTextFormatter	  300000	      3883 ns/op	  22.40 MB/s
BenchmarkLargeTextFormatter	  100000	     14305 ns/op	  20.06 MB/s
BenchmarkSmallColoredTextFormatter	  300000	      5114 ns/op	  26.79 MB/s
BenchmarkLargeColoredTextFormatter	   50000	     26918 ns/op	  20.54 MB/s
```