@noxiouz 2015-06-27
if it's ok, I can apply the same optimization to the colored formatter.