@aybabtme 2015-06-28
The idea LGTM, try with the switch I suggest, I guess avoiding the type case will squeeze out a bit more performance.