@noxiouz 2015-06-28
@aybabtme updated. The result is improved (+ ~8%):
```
BenchmarkSmallTextFormatter	  500000	      3540 ns/op	  24.57 MB/s
BenchmarkLargeTextFormatter	  100000	     13763 ns/op	  20.85 MB/s
```